Move `DevInterfaceData` initialization in its module
Before the initialization was made inside the `DevInterfaceSet::enumerate`
method. Moving this code to the `DevInterfaceData::fetch` method allows
for better handling of the safety invariants on the `DevInterfaceData`.

Other changes include not zero-initializing the `SP_DEVICE_INTERFACE_DATA`
and instead using `MaybeUninit` paired with `addr_of_mut!` to initialize
only the `cbSize` field.