Rework `DevInterfaceData`
This commit introduces the `utf16string` crate to store the string data
returned by the windows API functions without costly convertions to
other character encodings (like UTF-8, which was used before this commit)

The usage of `addr_of_mut` has been extended to other methods, to avoid
needless zero-initializations of winapi structs.

A new function is added, `alloc_slice_with_align` to handle the
allocation of variably sized structs, which it's used to allocate
the `SP_DEVICE_INTERFACE_DETAIL_DATA_W` used to retrieve the device
interface path.

The documentation is improved, most importantly lot of attention was
given to the safety comments over unsafe code.

Note that this commit doesn't change any code related to fetching the
properties values, as that part needs a lot more changes which will be
done in the following commits