Move and rework `DevInterfaceData` properties values fetching
This commit moves all the code related to fetching the device interface
property values to the `devdata::properties` submodule to better
separate the job of each module.

The properties value fetching is changed drastically:
the code has been organized in a way that allows the unsafe code to
better be documented and to allow for some optimizations for reducing
the number of allocations needed.
Now when the property has a known constant size, its value is allocated
on the stack. Only when the size is variable an allocation is made.