Switch from winapi to windows_sys
The windows_sys crate is offically supported by Microsoft so
it's a better alternative.
This commit replaces all references to the winapi types and
functions to the corresponding ones in windows_sys.

This commit also moves the win::Error consts generation to
the a build script in order to reduce compilation times.